+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Behavioural Changes and Deprecations
+
+- Removed the `EnableMetrics` option, following the .NET SDK where metrics are now always enabled and the option is a no-op. To drop metrics, use `SetBeforeSendMetric` and return `null`. ([#2828](https://github.com/getsentry/sentry-unity/pull/2828))
+
 ### Fixes
 
 - IL2CPP line numbers now work on Android x86/x86_64 builds. il2cpp fails to report the image UUID there, so the SDK falls back to looking the debug image up by name ([#2817](https://github.com/getsentry/sentry-unity/pull/2817))